+++ CKDownloadManager/DownloadViewController/CKDownloadProgress.h
@@ -11,5 +11,5 @@
 @interface CKDownloadProgress : UIView
 @property(nonatomic,strong) UIColor * progressColor;
 
--(void) setProgress:(float) progress animated:(BOOL)animated;
+-(void) setProgress:(CGFloat) progress animated:(BOOL)animated;
 @end
